requirements.txt
streamlit==1.37.1
pandas==2.3.2
numpy>=2.0,<3.0
python-dateutil==2.9.0.post0
